+++ packages/react-navigation/src/views/Header/HeaderTitle.js
@@ -17,7 +17,6 @@ const styles = StyleSheet.create({
     fontSize: Platform.OS === 'ios' ? 17 : 20,
     fontWeight: Platform.OS === 'ios' ? '700' : '500',
     color: 'rgba(0, 0, 0, .9)',
-    textAlign: Platform.OS === 'ios' ? 'center' : 'left',
     marginHorizontal: 16,
   },
 });